Extract smoke-all fixture helpers; record the parity spike (bd-xa4vv9tt)
Move fixture discovery, front-matter/tests-block reading, project-root
discovery, VFS population, user-grammar handles and the WASM loader out of
smokeAll.wasm.test.ts into hub-client/src/test-utils/smokeAllFixtures.ts so a
second smoke-all runner can share the exact fixture -> VFS semantics
(pass/skip counts unchanged: 131/131). Widen vitest.wasm.config.ts to
`*.wasm.test.{ts,tsx}` for a JSX runner.

The spike (claude-notes/research/2026-08-24-preview-render-parity-spike.md)
confirms the WASM module initialises under jsdom, both render_page_in_project
and render_page_for_preview succeed on plain fixtures, and a bare <Ast>
mounts with no shims. It found one rule the design table lacked (React's
RawBlock needs a host <div>) and four real divergences (bd-tmb2u5yu,
bd-294mbrcx, bd-q88zinyv, bd-qzwlhrlv).

Plan: claude-notes/plans/2026-08-24-preview-render-dom-parity-harness.md
